refactor-ir-cake: `staged` package removed
