+++ sc/shared/src/main/scala/sigma/compiler/Transforming.scala
@@ -1,7 +1,6 @@
-package sigma.compiler.staged
+package sigma.compiler
 
 import debox.{cfor, Buffer => DBuffer}
-import sigma.compiler.IRContext
 import sigma.data.{Lazy, Nullable}
 import sigma.reflection.RMethod
